修复 create-vtj  .gitignore 文件找不到错误

+++ create-vtj/src/generator.ts
@@ -1,6 +1,12 @@
 import { join } from 'path';
 import { readFileSync } from 'node:fs';
-import { copySync, readJsonSync, writeJsonSync, outputFile } from '@vtj/node';
+import {
+  copySync,
+  readJsonSync,
+  writeJsonSync,
+  outputFile,
+  pathExistsSync
+} from '@vtj/node';
 export interface IGeneratorOptions {
   root: string;
   template: string;
@@ -12,10 +18,13 @@ export function createLibrary(options: IGeneratorOptions) {
   const pkg = readJsonSync(join(options.root, 'package.json'));
   pkg.name = options.name;
   writeJsonSync(join(options.root, 'package.json'), pkg, { spaces: 2 });
-  outputFile(
-    join(options.root, '.gitignore'),
-    readFileSync(join(options.template, '.gitignore'), 'utf-8')
-  );
+  const gitignore = join(options.template, '.gitignore');
+  if (pathExistsSync(gitignore)) {
+    outputFile(
+      join(options.root, '.gitignore'),
+      readFileSync(gitignore, 'utf-8')
+    );
+  }
 }
 
 export function createApp(options: IGeneratorOptions) {
@@ -23,8 +32,11 @@ export function createApp(options: IGeneratorOptions) {
   const pkg = readJsonSync(join(options.root, 'package.json'));
   pkg.name = options.name;
   writeJsonSync(join(options.root, 'package.json'), pkg, { spaces: 2 });
-  outputFile(
-    join(options.root, '.gitignore'),
-    readFileSync(join(options.template, '.gitignore'), 'utf-8')
-  );
+  const gitignore = join(options.template, '.gitignore');
+  if (pathExistsSync(gitignore)) {
+    outputFile(
+      join(options.root, '.gitignore'),
+      readFileSync(gitignore, 'utf-8')
+    );
+  }
 }
